Tidy things up for a data fetcher interface.
This is a small tidy-up to move a bunch of Jinja2 specific code into a module. The idea is that we can start separating the parts of the code which are specific to PostgreSQL from the general parts using the `DataFetcher` interface. Then we can implement additional things conforming to `DataFetcher`, such as RAWR tiles.
