+++ tilequeue/query/__init__.py
@@ -1,0 +1,3 @@
+import tilequeue.query.postgres
+
+make_db_data_fetcher = postgres.make_db_data_fetcher
